refactor fst functions
